+++ fabric/changelog.md
@@ -1,1 +1,1 @@
-- Updated voicechat API
+- Added Opus encoder modes to voice chat API